requirements.txt
llama-cpp-python
pypdf
psutil
smbus2
pimoroni-bme280
